refactor: split grid_esp32_usb into ACM, HID, MIDI submodules
Split the monolithic grid_esp32_usb.c into separate files by USB class:
- grid_esp32_usb_acm.c/h  CDC-ACM serial, migrated from tinyusb_cdcacm
                           high-level API to native tud_cdc_* callbacks
- grid_esp32_usb_hid.c/h  HID (keyboard, mouse, gamepad)
- grid_esp32_usb_midi.c/h MIDI RX/TX

grid_esp32_usb.c now contains only USB descriptors and init.
grid_esp32_usb.h is an umbrella header that includes all subsystems.

The ACM migration to native callbacks removes the esp_tinyusb cdcacm
layer while keeping identical behaviour: tud_cdc_line_state_cb sets
usb_tx_ready on DTR, tud_cdc_tx_complete_cb clears the busy flag.

Also removes the stale #include "tinyusb_cdc_acm.h" from grid_esp32s3.c.
