+++ esp32s3/components/grid_esp32_usb/CMakeLists.txt
@@ -1,6 +1,9 @@
 idf_component_register(
     SRCS
         "grid_esp32_usb.c"
+        "grid_esp32_usb_acm.c"
+        "grid_esp32_usb_hid.c"
+        "grid_esp32_usb_midi.c"
     INCLUDE_DIRS
         "."
     REQUIRES
