+++ esp32s3/components/grid_esp32_usb/grid_esp32_usb_acm.h
@@ -1,0 +1,23 @@
+/*
+ * SPDX-FileCopyrightText: 2021-2022 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#pragma once
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+// Initialize ACM subsystem (called from grid_esp32_usb_init)
+void grid_esp32_usb_acm_init(void);
+
+// Platform API
+int32_t grid_platform_usb_serial_ready(void);
+int32_t grid_platform_usb_serial_write(char* buffer, uint32_t length);
+
+#ifdef __cplusplus
+}
+#endif
